requirements.txt
streamlit
openai
requests
beautifulsoup4
readability-lxml
markdown
python-docx
textract
weasyprint
